+++ nemo_nowcast/workers/awaken.py
@@ -21,7 +21,6 @@ sleep example worker finishes successfully.
 """
 import logging
 
-import nemo_nowcast.lib
 from nemo_nowcast.worker import NowcastWorker